requirements.txt
uvicorn==0.23.2
transformers==4.34.1
python-decouple==3.8
chardet==5.2.0
fastapi==0.104.0
accelerate==0.23.0